docs(PPDSC-2595): update homepage cms
