+++ site/pages/index.tsx
@@ -28,6 +28,13 @@ const GRID_SECTION_OVERRIDES: GridLayoutProps['overrides'] = {
   },
 };
 
+const heroCardFallbackContent = {
+  hero_card_title: `Latest blog`,
+  hero_card_description: `How an audio player component tells the story of NewsKit Design System's changing strategy.`,
+  hero_card_link_text: `Read on Medium`,
+  hero_card_link: `https://medium.com/newskit-design-system/how-an-audio-player-component-tells-the-story-of-newskit-design-systems-changing-strategy-8dc99d37ed67`,
+};
+
 const Index = ({
   releases,
   content,
@@ -52,11 +59,8 @@ const Index = ({
           }}
         >
           <FeatureCard
-            title={getValueFromCMS(content, 'hero_card_title') || 'Latest blog'}
-            description={
-              getValueFromCMS(content, 'hero_card_description') ||
-              "How an audio player component tells the story of NewsKit Design System's changing strategy"
-            }
+            title={getValueFromCMS(content, 'hero_card_title')}
+            description={getValueFromCMS(content, 'hero_card_description')}
             stylePrefix="worldDesignSystemsWeekCard"
             layout="horizontal"
             overrides={{
@@ -64,14 +68,8 @@ const Index = ({
               description: {typographyPreset: 'editorialSubheadline010'},
             }}
             buttonIcon={<IconFilledLaunch />}
-            buttonLabel={
-              getValueFromCMS(content, 'hero_card_link_text') ||
-              'Read on Medium'
-            }
-            buttonHref={
-              getValueFromCMS(content, 'hero_card_link') ||
-              'https://medium.com/newskit-design-system/how-an-audio-player-component-tells-the-story-of-newskit-design-systems-changing-strategy-8dc99d37ed67'
-            }
+            buttonLabel={getValueFromCMS(content, 'hero_card_link_text')}
+            buttonHref={getValueFromCMS(content, 'hero_card_link')}
             buttonOverrides={{
               paddingInline: 'space000',
               typographyPreset: 'utilityButton020',
@@ -138,6 +136,6 @@ export async function getStaticProps() {
     getSheets('Homepage'),
   ]);
 
-  const content = formatSheetData(data);
+  const content = {...heroCardFallbackContent, ...formatSheetData(data)};
   return {props: {releases, content}};
 }
